src/requirements.txt
boto3>=1.9.74
requests
python-dateutil